+++ src/app/modules/graphical-charter/graphical-charter-routing.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { HomeComponent } from './pages/home/home.component';
+import { CookingResolver } from '../cooking-recipe/resolvers/cooking.resolver';
 
 const routes: Routes = [
   {